Moved animator into seperate animators

+++ Scrile/Extensions/UIFontExtension.swift
@@ -13,11 +13,11 @@ extension UIFont {
     static let vagThin = "VAGRoundedStd-Thin"
     static let vagBold = "VAGRoundedStd-Bold"
     
-    static func vagBold(size: CGFloat) -> UIFont {
-        return UIFont(name: vagBold, size: size)!
+    static func vagBold(size: Double) -> UIFont {
+        return UIFont(name: vagBold, size: CGFloat(size))!
     }
     
-    static func vagThin(size: CGFloat) -> UIFont {
-        return UIFont(name: vagThin, size: size)!
+    static func vagThin(size: Double) -> UIFont {
+        return UIFont(name: vagThin, size: CGFloat(size))!
     }
 }
